Use FreeSql instead of PetaPoco.

+++ project/Elight.Web/Areas/System/Controllers/ItemController.cs
@@ -30,7 +30,7 @@ namespace Elight.Web.Areas.System.Controllers
         }
 
         [HttpPost, AuthorizeChecked]
-        public ActionResult Index(long pageIndex, long pageSize, string keyWord)
+        public ActionResult Index(int pageIndex, int pageSize, string keyWord)
         {
             var pageData = _itemService.GetList(pageIndex, pageSize, keyWord);
             var result = new LayPadding<Sys_Item>()
